+++ .gitignore
@@ -1,5 +1,5 @@
 .DS_Store
 coverage
-node_modules/
+node_modules
 npm-debug.log
-out
+lib
