--- index.js
@@ -1,8 +1,0 @@
-'use strict'
-
-module.exports = {
-  Server: require('./lib/server'),
-  // Errors:
-  TaskError: require('./lib/errors/task-error'),
-  TaskFatalError: require('./lib/errors/task-fatal-error')
-}
