+++ src/rabbitmq.js
