+++ src/worker.js
