+++ test/functional/timeout.js
@@ -6,13 +6,13 @@ const sinon = require('sinon')
 const assert = chai.assert
 
 // Ponos Tooling
-const ponos = require('../../')
+const ponos = require('../../src')
 const TimeoutError = require('bluebird').TimeoutError
 const testWorker = require('./fixtures/timeout-worker')
 const testWorkerEmitter = testWorker.emitter
 
 // require the Worker class so we can verify the task is running
-const _Worker = require('../../lib/worker')
+const _Worker = require('../../src/worker')
 // require the error module so we can see the error printed
 const _Bunyan = require('bunyan')
